Handle multiple error types!
Had to build my own enum type that wrapped two error types and
implemented the From trait for them, before realizing that I was using
the .collect() superfish wrong and that this basically is what the Box
type does. So .collect transforms to a Result with the original error,
and `?` transforms to the Box-of-thing-with-the-Error-trait.